+++ package.json
@@ -59,6 +59,7 @@
     "ts-node": "10.9.1",
     "typescript": "5.0.4",
     "webpack": "5.82.0",
+    "webpack-bundle-analyzer": "^4.8.0",
     "webpack-cli": "5.0.2"
   }
 }
