+++ src/modules/bundle/index.ts
@@ -1,0 +1,1 @@
+export { BundleManager } from './BundleManager'
