+++ src/modules/cli/cli.ts
@@ -1,0 +1,5 @@
+export class Cli {
+    constructor() {
+        //
+    }
+}
