+++ src/modules/config/index.ts
@@ -1,0 +1,1 @@
+export { Config } from './Config'
