+++ src/modules/json-rpc-handler/index.ts
@@ -1,0 +1,1 @@
+export { RpcMethodHandler } from './RpcMethodHandler'
