+++ src/modules/json-rpc-handler/services/Web3.ts
@@ -1,4 +1,4 @@
-import Config from '../../Config'
+import { Config } from '../../config'
 
 export class Web3API {
