+++ src/modules/json-rpc-handler/services/index.ts
@@ -1,0 +1,3 @@
+export { EthAPI } from './Eth'
+export { DebugAPI } from './Debug'
+export { Web3API } from './Web3'
