+++ src/modules/json-rpc-server/index.ts
@@ -1,0 +1,1 @@
+export { JsonrpcHttpServer } from './JsonRpcHttpServer'
