+++ src/modules/provider/ProviderService.ts
@@ -1,4 +1,4 @@
-import Config from './Config'
+import { Config } from '../config'
 
 export class ProviderService {
     async getNetwork() {
