+++ src/modules/provider/index.ts
@@ -1,0 +1,1 @@
+export { ProviderService } from './ProviderService'
