+++ src/modules/types/mempool.types.ts
@@ -1,0 +1,6 @@
+import { UserOperation } from './userop.types'
+
+export interface MempoolEntry {
+    userOp: UserOperation;
+    userOpHash: string;
+}
