+++ src/modules/types/rcp.types.ts
@@ -1,0 +1,6 @@
+export interface JsonRpcRequest {
+    jsonrpc: '2.0';
+    method: string;
+    params: any[];
+    id: number | string;
+}
